Implemented an object pool.
Implemented the object pool pattern with a free list (that cannibalizes
particle's memory, so as not to use another entire buffer -> waste of
memory/space).

Still a lot of cleaning up to do for the math and alignment.

Times in DEBUG improved from the following:

update is 60-100
draw is 77-84

to this:

update is 23-36
draw is 16-30